requirements.txt
django-rest-swagger==2.2.0
Django==4.1.4
djangorestframework==3.14.0
